pkp-pkp-lib#1131 CSRF checking for confirmation modal actions

+++ controllers/grid/issueGalleys/IssueGalleyGridHandler.inc.php
@@ -252,9 +252,8 @@ class IssueGalleyGridHandler extends GridHandler {
 		if ($issueGalleyForm->validate($request)) {
 			$issueId = $issueGalleyForm->execute($request);
 			return DAO::getDataChangedEvent($issueId);
-		} else {
-			return new JSONMessage(false);
 		}
+		return new JSONMessage(false);
 	}
 
 	/**
@@ -265,9 +264,11 @@ class IssueGalleyGridHandler extends GridHandler {
 	function delete($args, $request) {
 		$issueGalleyDao = DAORegistry::getDAO('IssueGalleyDAO');
 		$issueGalley = $this->getAuthorizedContextObject(ASSOC_TYPE_ISSUE_GALLEY);
-		$issueGalley->getId();
-		$issueGalleyDao->deleteObject($issueGalley);
-		return DAO::getDataChangedEvent();
+		if ($issueGalley && $request->checkCSRF()) {
+			$issueGalleyDao->deleteObject($issueGalley);
+			return DAO::getDataChangedEvent();
+		}
+		return new JSONMessage(false);
 	}
 
 	/**
